pyproject.toml
[project]
name = "nexus-bot-ai"
version = "0.1.0"
description = "Universal Discord Bot Template"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.5.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.30.0",
    "python-dotenv>=1.0.0",
    "apscheduler>=3.10.0",
    "aiohttp>=3.9.0",
]

[tool.ruff]
line-length = 88
target-version = "py313"
